# lib/api-zod/src/generated/types/changePasswordInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kampulse Recruitment & Employee Onboarding Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ChangePasswordInput {
  currentPassword: string;
  newPassword: string;
}